pyproject.toml
[tool.poetry]
name = "langchainexplorations"
version = "0.1.0"
description = ""
authors = ["Giuseppe Rosa <giusepper11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.350"
openai = "^0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
